feat(rule): require parens for arrow functions
## Reasoning:

There are four possible ways to use an arrow function with a single argument:
1. Parens, implicit return
2. No-parens, implicit return
3. Parens, function block return
4. No-parens, function block return

When dealing with mutiple arguments parens are required, which brings additional complexity to the various ways of writing an arrow function.

No-parens with implicit return adds additional cognitive load since it requires you to read past the argument to the arrow to discover it is a function:

```js
doSomething(fooBarBaz => fooBarBaz);
```

Removing the option for paren-less arrow functions reduces possible configurations to just 2, establishes a more consistent pattern for recognizing functions, and consequently makes the code clearer and more readable.